feat: change quantidadeRealizada type to BigDecimal for improved precision

+++ src/main/java/com/branches/atividade/dto/request/CreateAtividadeDeRelatorioRequest.java
@@ -15,7 +15,7 @@ import java.util.List;
 public record CreateAtividadeDeRelatorioRequest(
         @NotBlank(message = "O campo 'descricao' é obrigatório")
         String descricao,
-        Integer quantidadeRealizada,
+        BigDecimal quantidadeRealizada,
         String unidadeMedida,
         @DecimalMax(value = "100.00", message = "O campo 'porcentagemConcluida' deve ser no máximo 100.00")
         @DecimalMin(value = "0.00", message = "O campo 'porcentagemConcluida' deve ser no mínimo 0.00")
